Moved PreviewPane to a separate directory. Extracted Preview component to a separate file.
