+++ src/components/PreviewPane/PreviewPane.js
@@ -1,47 +1,10 @@
 import React, { PropTypes } from 'react';
 import { render } from 'react-dom';
 import ImmutablePropTypes from 'react-immutable-proptypes';
-import registry from '../lib/registry';
-import { resolveWidget } from './Widgets';
+import registry from '../../lib/registry';
+import { resolveWidget } from '../Widgets';
+import Preview from './Preview';
 import styles from './PreviewPane.css';
-
-class Preview extends React.Component {
-
-  previewFor(field) {
-    const { entry, getMedia } = this.props;
-    const widget = resolveWidget(field.get('widget'));
-    return React.createElement(widget.preview, {
-      field: field,
-      value: entry.getIn(['data', field.get('name')]),
-      getMedia: getMedia,
-    });
-  }
-
-  render() {
-    const { collection } = this.props;
-    if (!collection) {
-      return null;
-    }
-
-    return <div>
-      {
-        collection.get('fields').map(field => (
-          <div
-            key={field.get('name')}
-          >
-            {this.previewFor(field)}
-          </div>
-        ))
-      }
-    </div>;
-  }
-}
-
-Preview.propTypes = {
-  collection: ImmutablePropTypes.map.isRequired,
-  entry: ImmutablePropTypes.map.isRequired,
-  getMedia: PropTypes.func.isRequired,
-};
 
 export default class PreviewPane extends React.Component {
